Fix e0c420a96b9f70dce7d2087f49fdad3cb1dae3b3
If options.groupsize was not specified (and thus zero), we would use
zero as local work size, which resulted in an infinite loop on some
devices.

Fix by checking for options.groupsize to be set before using it. Also
move the calculation of the second kernel launch size to before the
loop, since it's fixed anyway.